requirements.txt
absl-py==1.0.0
addict==2.4.0
affine==2.3.1
albumentations==1.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
boto3==1.15.18
botocore==1.18.18
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.1.0
commonmark==0.9.1
coremltools==5.2.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
easydict==1.9
efficientnet-pytorch==0.6.3
entrypoints==0.4
executing==0.8.2
fast_ctc_decode==0.3.2
fastapi==0.75.0
filelock==3.7.1
flatbuffers==1.12
fonttools==4.28.5
gast==0.4.0
gluonnlp==0.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
huggingface-hub==0.0.12
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
imutils==0.5.4
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.16.0
jstyleson==0.0.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.0
jupyter-core==4.11.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
libclang==13.0.0
lmdb==1.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==2.0.4
mmcls==0.23.1
mmcv==1.5.2
model-index==0.1.11
mpmath==1.2.1
munch==2.5.0
mxnet==1.7.0.post2
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.6.3
nibabel==3.2.2
nltk==3.7
notebook==6.4.12
numpy==1.23.3
onnx==1.8.1
onnxoptimizer==0.2.7
onnxruntime==1.8.0
opencv-contrib-python==4.5.5.64
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openvino==2022.1.0
openvino-dev==2022.1.0
openvino-telemetry==2022.1.1
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parasail==1.2.4
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.4.1
pretrainedmodels==0.7.4
prettytable==3.3.0
progress==1.6
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.1
py-cpuinfo==8.0.0
pyclipper==1.3.0.post2
pycparser==2.21
pydantic==1.9.0
pydicom==2.3.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.0
qudida==0.0.4
rasterio==1.3.2
rawpy==0.17.1
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
s3transfer==0.3.7
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
Shapely==1.8.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.1.3
starlette==0.17.1
submitit==1.4.2
subprocess32==3.5.4
sympy==1.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflowjs==3.18.0
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.5.0
texttable==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2022.8.12
timm==0.4.12
tinycss2==1.1.1
tokenizers==0.10.3
tomli==1.2.3
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
transformers==4.8.1
typing_extensions==4.3.0
urllib3==1.25.11
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==2.1.1
widgetsnbextension==3.5.2
wrapt==1.14.0
yacs==0.1.8
yapf==0.32.0
yaspin==2.1.0
zipp==3.8.1
